Add files via upload
This is my first full-fledged game written in Java. There is a lot to optimize and reduce the number of lines and commands, but at the moment it is very difficult for me to do it. In the future, I want to simplify this game in order to optimize and hone my work skills.